refactor pt 2 => separate component for panels
TODO
- Adjust css from tabspanelcomponent and tabs page
- Open model option should open in the chosen panel
- Finish refactor
- Unit tests
