+++ COMETwebapp/Components/Tabs/TabsPanelComponent.razor.css
@@ -18,7 +18,6 @@
     justify-content: center;
     align-items: center;
     height: 90vh;
-    width: 100%;
 }
 
 #tabs-page-content {
@@ -29,6 +28,6 @@
     overflow-x: auto;
 }
 
-.panel-view:not(:first-child) {
-    border-left: solid 1px var(--colors-gray-100);
-}
+    .panel-view:not(:first-child) {
+        border-left: solid 1px var(--colors-gray-100);
+    }
